+++ LibManSys-Backend/src/main/java/ca/mcgill/ecse321/projectgroup09/dao/BookRepository.java
@@ -1,11 +1,14 @@
 package ca.mcgill.ecse321.projectgroup09.dao;
 
 import org.springframework.data.repository.CrudRepository;
-
-import ca.mcgill.ecse321.projectgroup09.models.*;
-
-public interface BookRepository extends CrudRepository<Book, Long> {
+import ca.mcgill.ecse321.projectgroup09.models.Book;
+/**
+ * 
+ * @author Zarif Ashraf
+ *
+ */
+public interface BookRepository extends CrudRepository<Book, Integer> {
 	
-	Book findBookBylibraryItemID(Long libraryItemID);
+	Book findBookBylibraryItemID(Integer libraryItemID);
 	Book findBookByisbn(String isbn);
 }
