+++ LibManSys-Backend/src/main/java/ca/mcgill/ecse321/projectgroup09/dao/LibraryItemRepository.java
@@ -1,14 +1,18 @@
 package ca.mcgill.ecse321.projectgroup09.dao;
 
 import org.springframework.data.repository.CrudRepository;
-
-import ca.mcgill.ecse321.projectgroup09.models.*;
+import ca.mcgill.ecse321.projectgroup09.models.LibraryItem;
+import ca.mcgill.ecse321.projectgroup09.models.Loan;
 import ca.mcgill.ecse321.projectgroup09.models.LibraryItem.ItemStatus;
 import java.util.List;
-
-public interface LibraryItemRepository extends CrudRepository<LibraryItem, Long> {
+/**
+ * 
+ * @author Zarif Ashraf
+ *
+ */
+public interface LibraryItemRepository extends CrudRepository<LibraryItem, Integer> {
 	
-	LibraryItem findLibraryItemByLibraryItemID(Long libraryItemID);
+	LibraryItem findLibraryItemByLibraryItemID(Integer libraryItemID);
 	LibraryItem findLibraryItemByLoan(Loan loan);
 	List<LibraryItem> findLibraryItemByItemStatus(ItemStatus itemstatus);
 	List<LibraryItem> findLibraryItemByTitle(String title);
