+++ LibManSys-Backend/src/main/java/ca/mcgill/ecse321/projectgroup09/dao/MemberRepository.java
@@ -5,10 +5,14 @@ import ca.mcgill.ecse321.projectgroup09.models.Member;
 import ca.mcgill.ecse321.projectgroup09.models.Booking;
 import ca.mcgill.ecse321.projectgroup09.models.Loan;
 import java.util.List;
-
-public interface MemberRepository extends CrudRepository<Member, Long> {
+/**
+ * 
+ * @author Zarif Ashraf
+ *
+ */
+public interface MemberRepository extends CrudRepository<Member, Integer> {
 	
-	Member findMemberByLibCardNumber(Long libCardNumber);
+	Member findMemberByLibCardNumber(Integer libCardNumber);
 	Member findMemberByBooking(Booking booking);
 	Member findMemberByLoan(Loan loan);
 	List<Member> findMemberByVerificationStatus(boolean isVerified);
