+++ LibManSys-Backend/src/main/java/ca/mcgill/ecse321/projectgroup09/dao/NewspaperRepository.java
@@ -3,10 +3,14 @@ package ca.mcgill.ecse321.projectgroup09.dao;
 import org.springframework.data.repository.CrudRepository;
 import ca.mcgill.ecse321.projectgroup09.models.Newspaper;
 import java.util.List;
-
-public interface NewspaperRepository extends CrudRepository<Newspaper, Long> {
+/**
+ * 
+ * @author Zarif Ashraf
+ *
+ */
+public interface NewspaperRepository extends CrudRepository<Newspaper, Integer> {
 	
-	Newspaper findNewspaperBylibraryItemID(Long libraryItemID);
+	Newspaper findNewspaperBylibraryItemID(Integer libraryItemID);
 	List<Newspaper> findNewspapersByJournalName(String journalName);
 	List<Newspaper> findNewspapersByChiefEditor(String chiefEditor);
 }
